Created API Blueprint for ML model endpoint. Need to fix actual prediction ability.
