+++ app/ml_model_api/__init__.py
@@ -1,0 +1,1 @@
+from ml_model_api.ml_model_api_blueprint import ml_model_api_blueprint
